feat(etapa-5/fatia3): Q5b - preserva quantity fracionaria no codigo (mapper+RPC)
quoteMoney.ts: coerceQuantity vira roundQuantity (quantiza a 3 casas, delega a
roundMoney) - nao arredonda mais a inteiro, reflete quote_items.quantity numeric
(20260719001350). isFractionalQuantity mantido (agora informativo, nao mais
lossy).

quoteMapper.ts: mapLocalQuoteItemToSupabaseItem usa roundQuantity em vez de
coerceQuantity.

RPC import_quote_with_items (20260719001400): cast do item na SELECT/INSERT
dos filhos de ::integer para ::numeric; comentario de exemplo atualizado
(quantity fracionaria); pre-requisito no cabecalho passa a incluir 20260719001350.

Testes: roundQuantity substitui os casos de coerceQuantity (preserva fracao,
quantiza a 3 casas); quoteMapper ganha 2 casos novos (preserva 1.5 sem
arredondar; quantiza 1.23456 -> 1.235) + regressao de quantidade inteira.

Gates: tsc 0, suite 118/118 (subiu de 116), lint 89/68 sem regressao.

+++ supabase/migrations/20260719001400_etapa5_fatia3_import_quote_with_items_rpc.sql
@@ -79,9 +79,13 @@
 --
 -- Transacional (CREATE FUNCTION é DDL simples) — não toca dado existente, seguro
 -- aplicar mesmo com quotes populada.
--- >>> PRÉ-REQUISITO: as migrations 20260719001000 a 20260719001300 já aplicadas
+-- >>> PRÉ-REQUISITO: as migrations 20260719001000 a 20260719001350 já aplicadas
 --     (colunas client_id/opportunity_id/source_local_id + UNIQUE
---     ux_quotes_source_local, que é o arbiter do ON CONFLICT abaixo).
+--     ux_quotes_source_local, que é o arbiter do ON CONFLICT abaixo + Q5b,
+--     quote_items.quantity numeric, para o cast ::numeric abaixo preservar
+--     fração — sem ela, o cast ainda funciona mas o cast de atribuição
+--     numeric->integer no INSERT ARREDONDA (não trunca) ao gravar numa coluna
+--     ainda integer, perdendo a fração mesmo assim).
 -- ============================================================================
 
 CREATE OR REPLACE FUNCTION public.import_quote_with_items(
@@ -98,7 +102,8 @@ CREATE OR REPLACE FUNCTION public.import_quote_with_items(
   p_total numeric,
   p_status text,
   p_archived boolean,
-  -- Array de itens: [{ "name": "...", "quantity": 1, "unit_price": 10.5, "service_id": null }, ...]
+  -- Array de itens: [{ "name": "...", "quantity": 1.5, "unit_price": 10.5, "service_id": null }, ...]
+  -- quantity é numeric (Q5b, 20260719001350) — aceita fração (ex.: horas).
   p_items jsonb
 )
 RETURNS public.quotes
@@ -160,7 +165,7 @@ BEGIN
       v_quote.id,
       NULLIF(item->>'service_id', '')::uuid,
       item->>'name',
-      (item->>'quantity')::integer,
+      (item->>'quantity')::numeric,
       (item->>'unit_price')::numeric
     FROM jsonb_array_elements(p_items) AS item;
   END IF;
